Shanyrak/app/config.py
import os
from dotenv import load_dotenv


load_dotenv()

class Settings:
    PROJECT_NAME: str = "Shanyraq.kz"
    DEBUG: bool = os.getenv("DEBUG", "True") == "True"
    SECRET_KEY: str = os.getenv("SECRET_KEY", "super-secret-key")
    ALGORITHM: str = "HS256"
    ACCESS_TOKEN_EXPIRE_MINUTES: int = 60 * 24
    DATABASE_URL: str = os.getenv("DATABASE_URL", "sqlite:///./shanyraq.db")

settings = Settings()
