Shanyrak/pyproject.toml
[project]
name = "shanyrak"
version = "0.1.0"
description = ""
authors = [
    {name = "nurlibaj",email = "nurlibaj@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pydantic (>=2.11.1,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "alembic (>=1.15.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
